test-data-seeder/requirements.txt
flask>=3.0.0
supabase>=2.5.0
python-dotenv>=1.0.0
faker>=22.0.0
playwright>=1.40.0
# pandas is needed by validate_analytics.py smoke tests (analytics modules)
pandas>=2.0.0
